Removing dependencies in use cases
