+++ apps/project-manager-api/src/domain/use-cases/use-cases.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { ProjectsModule } from './projects/projects.module';
-import { TasksModule } from '../../../../tasks/src/domain/use-cases/tasks/tasks.module';
+import { TasksModule } from '../../../../tasks/src/domain/use-cases/tasks.module';
 import { UsersModule } from './users/users.module';
 
 @Module({
