+++ apps/tasks/src/domain/domain.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { TasksModule } from './use-cases/tasks/tasks.module';
+import { TasksModule } from './use-cases/tasks.module';
 @Module({
   imports: [TasksModule],
   exports: [TasksModule],
