+++ apps/tasks/src/domain/use-cases/get-all-tasks.service.ts
@@ -1,20 +1,14 @@
 import { Injectable } from '@nestjs/common';
 import { BaseUseCase } from '@project-manager-api/domain/use-cases/base-use-case';
-import { UsersRepositoryService } from '@project-manager-api/domain/repositories/users.repository.service';
-import { TasksRepositoryService } from '../../repositories/tasks.repository.service';
-import { ITask } from '../../interfaces/task.interface';
+import { TasksRepositoryService } from '../repositories/tasks.repository.service';
+import { ITask } from '../interfaces/task.interface';
 
 Injectable();
 export class GetAllTasksService implements BaseUseCase {
   constructor(
-    private readonly usersRepository: UsersRepositoryService,
     private readonly tasksRepository: TasksRepositoryService,
   ) {}
   async execute(payload: { userId: number }): Promise<ITask[]> {
-    const userData = await this.usersRepository.findById(payload.userId);
-    if (!userData) {
-      throw new Error('Usuário não encontrado');
-    }
     const tasks = await this.tasksRepository.findAll(payload.userId);
     if (!tasks) {
       throw new Error('Erro ao listar tarefas');
