+++ apps/tasks/src/domain/use-cases/tasks.module.ts
@@ -3,10 +3,10 @@ import { GetAllTasksService } from './get-all-tasks.service';
 import { GetTaskByIdService } from './get-task-by-id.service';
 import { CreateTaskService } from './create-task.service';
 import { UpdateTaskService } from './update-task.service';
-import { DatabaseModule } from '@project-manager-api/infrastructure/database/database.module';
+import { InfrastructureModule } from '../../infrastructure/infraestructure.module';
 
 @Module({
-  imports: [DatabaseModule],
+  imports: [InfrastructureModule],
   providers: [GetAllTasksService, GetTaskByIdService, CreateTaskService, UpdateTaskService],
   exports: [GetAllTasksService, GetTaskByIdService, CreateTaskService, UpdateTaskService],
 })
