+++ apps/tasks/src/gateways/controllers/dtos/create-task.dto.ts
@@ -4,8 +4,14 @@ export class CreateTaskDto {
   @IsNotEmpty({ message: 'O nome da tarefa precisa ser definido' })
   @IsString()
   name: string;
+
   @IsNotEmpty({ message: 'O status da tarefa precisa ser definido' })
   @IsString()
   status: 'pending' | 'completed';
+
   projectId: number;
+
+  userId: number;
+
+
 }
